backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
langchain>=0.1.0
langchain-openai>=0.1.0
pydantic>=2.6
inngest
